Agent: Move _filter_invalid_ranges to NetworkRanges
